Split AdminHandler god object into 7 domain-specific handlers
The monolithic AdminHandler (2,312 lines, 29+ methods, 11 constructor
params) has been split into focused handler structs, each taking only
the services it needs:

- AdminStatsHandler (1 dep) — dashboard stats + activity feed
- AdminShowHandler (5 deps) — pending shows, approve/reject, batch ops, import/export
- AdminVenueHandler (2 deps) — unverified venues, verify, pending edits
- AdminUserHandler (1 dep) — user listing
- AdminTokenHandler (1 dep) — API token CRUD
- AdminDataHandler (1 dep) — data sync export/import
- AdminDiscoveryHandler (1 dep) — discovery import/check

Shared helpers (getUserID, ptrString, parseDate) moved to helpers.go.
All route paths, HTTP methods, and handler logic are unchanged.
Tests updated to use domain-specific handler structs.

Closes PSY-188
